Decoupled the hygraph-ui and hygraph-dynamic-rows functionality so Hygraph CMS upgrades aren’t necessary
