--- examples/magento-graphcms/components/GraphCMS/DynamicRowRenderer.graphql
@@ -1,16 +1,0 @@
-fragment DynamicRowRenderer on DynamicRow @inject(into: ["DynamicRow"]) {
-  row {
-    ...RowColumnOne
-    # ...RowColumnTwo
-    # ...RowColumnThree
-    # ...RowBlogContent
-    # ...RowHeroBanner
-    # ...RowSpecialBanner
-    ...RowQuote
-    # ...RowButtonLinkList
-    # ...RowServiceOptions
-    # ...RowContentLinks
-    # ...RowProduct
-    ...RowLinks
-  }
-}
