+++ packages/hygraph-dynamic-rows/graphql/Conditions.graphql
