+++ packages/hygraph-dynamic-rows/graphql/DynamicRow.graphql
@@ -11,5 +11,9 @@ fragment DynamicRow on DynamicRow @injectable {
     ... on Node {
       id
     }
+
+    ...RowColumnOne
+    ...RowQuote
+    ...RowLinks
   }
 }
