+++ packages/hygraph-dynamic-rows/graphql/DynamicRows.graphql
