--- packages/hygraph-ui/components/PageContent/index.ts
@@ -1,1 +1,0 @@
-export * from './pageContent'
