+++ packages/hygraph-ui/components/index.ts
@@ -1,3 +1,2 @@
 export * from './RichText'
 export * from './Asset/Asset'
-export * from './PageContent'
