+++ packages/hygraph-ui/graphql/index.ts
@@ -1,7 +1,4 @@
 export * from './AllPageRoutes.gql'
-export * from './Conditions.gql'
-export * from './DynamicRow.gql'
-export * from './DynamicRows.gql'
 export * from './Page.gql'
 export * from './PageLink.gql'
 export * from './PagesContent.gql'
