+++ packages/hygraph-ui/index.ts
@@ -1,3 +1,4 @@
 export * from './components'
 export * from './graphql'
+export * from './lib'
 export * from './links/createHygraphLink'
