+++ packagesDev/next-config/__tests__/utils/resolveDependenciesSync.ts
@@ -9,12 +9,14 @@ it('resolves dependences', () => {
     Map {
       "." => "examples/magento-graphcms",
       "@graphcommerce/cli" => "packages/cli",
+      "@graphcommerce/hygraph-cli" => "packages/hygraph-cli",
       "@graphcommerce/demo-magento-graphcommerce" => "packages/demo-magento-graphcommerce",
       "@graphcommerce/googleanalytics" => "packages/googleanalytics",
       "@graphcommerce/magento-cart-shipping-method" => "packages/magento-cart-shipping-method",
       "@graphcommerce/googlerecaptcha" => "packages/googlerecaptcha",
       "@graphcommerce/googletagmanager" => "packages/googletagmanager",
-      "@graphcommerce/graphcms-ui" => "packages/graphcms-ui",
+      "@graphcommerce/hygraph-dynamic-rows" => "packages/hygraph-dynamic-rows",
+      "@graphcommerce/graphcms-ui" => "packages/hygraph-ui",
       "@graphcommerce/lingui-next" => "packages/lingui-next",
       "@graphcommerce/next-config" => "packagesDev/next-config",
       "@graphcommerce/magento-cart-billing-address" => "packages/magento-cart-billing-address",
